Handle 24 month terms
The new FeeBoundary object allows us to perform the same operations
regardless of whether the LoadApplication term was 12 or 24 months. The
term only needs to be checked when getting the upper and lower bounds.

The interpolate method has been split into interpolate and
perfromInterpolate. interpolate understands FeeBoundary and
LoanApplication objects while performInterpolate is pure math.